mini_data_pipeline/processing/curated_zone.py
"""Module: curated_zone.py
Handles writing Parquet to the curated zone and registering Athena tables."""
import boto3
import pyarrow.parquet as pq
import pyarrow as pa

def write_curated_parquet(bucket: str, enriched_key: str, curated_prefix: str, timestamp: str) -> str:
    """Read JSON, convert to Parquet, store under curated prefix.
    :return: S3 path of Parquet files"""
    import s3fs
    fs = s3fs.S3FileSystem()
    input_path = f's3://{bucket}/{enriched_key}'
    table = pq.read_table(input_path, filesystem=fs)
    out_prefix = f'{curated_prefix}{timestamp}/'
    out_path = f's3://{bucket}/{out_prefix}'
    pq.write_to_dataset(table, root_path=out_path, filesystem=fs, partition_cols=['year','month'])
    return out_path

def register_athena_table(bucket: str, parquet_path: str):
    """Use the Glue Data Catalog/ATHENA API to create or update an external table."""
    glue = boto3.client('glue')
    # Placeholder: implement create_table or update_table
    return
