mini_data_pipeline/requirements.txt
boto3
requests
zeep
grpcio
pyspark
pydeequ
great_expectations
psycopg2-binary
